Sparando controllers e criado endpoint getAllUsers
