+++ semana14/ToDoList/src/controllers/UsersController.ts
@@ -1,7 +1,5 @@
 import { Request, Response } from 'express';
 
-const { formatStringDate } = require('../utils/utils');
-
 //Configurações do banco
 import db from '../config/database';
 
@@ -57,37 +55,13 @@ const updateUser = async (
   }
 };
 
-const createTask = async (
-  title: string,
-  description: string,
-  limitDate: Date,
-  creatorUserId: string,
-): Promise<void> => {
+const getAllUsers = async (): Promise<any> => {
   try {
-    await db
-    .insert({
-      id: Date.now(),
-      title,
-      description,
-      limit_date: limitDate,
-      creator_user_id: creatorUserId
-    })
-    .into("TodoListTask ");    
-  } catch (error) {
-    console.log(error) 
-  }
-};
-
-const getTaskById = async (
-  id: string
-): Promise<any> => {
-  try {
-    const task = await db
+    const users = await db
     .select('*')
-    .from('TodoListTask')
-    .where({ id });
+    .from('TodoListUser');
   
-    return task[0];
+    return users;
     
   } catch (error) {
     console.log(error);
@@ -168,52 +142,14 @@ module.exports = {
       });
     }
   },
-  async createTask(req: Request, res: Response) {
+  async getAllUsers(req: Request, res: Response) {
     try {
-      //Validação todos os campos obrigatórios
-      const keys = Object.keys(req.body)
-  
-      for (const key of keys) {
-        if (req.body[key] == "")
-          return res.status(400).send({ message: "Por gentileza preencha todos os campos corretamente!"})
-      }
+      const users = await getAllUsers();
 
-    const { title, description, limitDate, creatorUserId } = req.body;
-
-    const limitDateFormart = await formatStringDate(limitDate);
-  
-    await createTask(
-        title,
-        description,
-        limitDateFormart,
-        creatorUserId
-      );
-  
-      res.status(200).send();
-    } catch (error) {
-      res.status(400).send({
-        message: error.message,
-      });
-    }
-  },
-  async getTaskById(req: Request, res: Response) {
-    try {
-      const taskId = req.params.id;
-
-      if (!taskId.length) {
-        return res.status(400).send({ message: "Informe um ID para pesquisa da tarefa"});
-      }
-
-      const task = await getTaskById(taskId);
-
-      if (Object.entries(task).length === 0) {
-        return res.status(400).send("Tarefa não encontrado para o ID informado!");
-      }
-
-      return res.status(200).send( { task });;
+      return res.status(200).send({ users });;
     } catch (error) {
       return res.status(400).send({
-        message: "Tarefa não encontrado para o ID informado!",
+        message: error.message,
       });
     }
   },
