+++ semana14/ToDoList/src/routes/router.ts
@@ -2,14 +2,17 @@ import express from 'express';
 const router = express.Router();
 
 //Controller
-const ToDoListController = require('../controllers/ToDoListControler');
+const UsersController = require('../controllers/UsersController');
+const TasksController = require('../controllers/TasksController');
 
+//Rotas usuários
+router.put("/user", UsersController.createUser);
+router.get("/user/:id", UsersController.getUserById);
+router.put("/user/edit/:id", UsersController.updateUser);
+router.get("/users/all", UsersController.getAllUsers);
 
-//Rotas
-router.put("/user", ToDoListController.createUser);
-router.get("/user/:id", ToDoListController.getUserById);
-router.put("/user/edit/:id", ToDoListController.updateUser);
-router.put("/task", ToDoListController.createTask);
-router.get("/task/:id", ToDoListController.getTaskById);
+//Rotas Tarefas
+router.put("/task", TasksController.createTask);
+router.get("/task/:id", TasksController.getTaskById);
 
 module.exports = router;
